pyproject.toml
[tool.poetry]
name = "company_challenge"
version = "0.0.1"
description = "My proposed solution to a company challenge."
authors = ["Felipe <feslima93@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
psycopg2-binary = "^2.9.3"
django-extensions = "^3.1.5"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
pytest = "^7.0.0"
black = "^22.1.0"
isort = "^5.10.1"
autoflake = "^1.4"
mypy = "^0.931"
Faker = "^12.3.0"
factory-boy = "^3.2.1"
pytest-django = "^4.5.2"
django-stubs = "^1.9.0"
djangorestframework-stubs = "^1.4.0"
pytest-mock = "^3.7.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py", "tests/*.py"]
filterwarnings = []

[tool.mypy]
python_version = 3.9
no_incremental = true
show_traceback = true
raise_exceptions = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "core.settings"

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "psycopg2.*",
    "django_stubs_ext.*",
    "django_extensions.*",
    "factory.*",
    "pytest_factoryboy.*",
]
ignore_missing_imports = true

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
